requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
Flask-SQLAlchemy==2.4.4
alembic==1.4.3
psycopg2==2.8.6
